Adjust transform indexing
When data is provided as 5D ([time, channel, z, y, x]) the application of the transform does not work given a matrix dimension mismatch (the matrix multiplication of scaling_xfm or translation_xfm with affine). This change fixes that by only taking the last 3 dimensions of the transform (which correspond to z,y,x).